pi/pi_requirements.txt
mysql-connector-python
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-circuitpython-led-animation
python-socketio[client]

